fix: use BigDecimal for monetary amounts to avoid floating-point precision errors
Replace double with BigDecimal in Transaction, Account and TransactionService.
Introduce nested AccountState helper class to hold mutable per-account aggregation state.

+++ src/main/java/com/example/microservice/transactions/model/Account.java
@@ -1,3 +1,5 @@
 package com.example.microservice.transactions.model;
 
-public record Account(String account, int debitCount, int creditCount, double balance) {}
+import java.math.BigDecimal;
+
+public record Account(String account, int debitCount, int creditCount, BigDecimal balance) {}
